requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
mcp==1.0.0
google-generativeai==0.8.5
